moved nav types into types folder
