+++ src/navigation/stacks/DraftStackNavigator.tsx
@@ -1,16 +1,18 @@
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import React from 'react';
 import DraftScreen from '../../screens/Drafting/Draft';
-import { DraftStackParamList } from '../types';
+import { DraftStackParamList } from '../../types/navigation';
 
 const DraftStack = createNativeStackNavigator<DraftStackParamList>();
 
 export default function DraftStackNavigator() {
-    return(
-        <DraftStack.Navigator screenOptions={{
-            headerShown: false,
-          }}>
-            <DraftStack.Screen name="DraftScreen" component= {DraftScreen} />
-        </DraftStack.Navigator>
-    )
+  return (
+    <DraftStack.Navigator
+      screenOptions={{
+        headerShown: false,
+      }}
+    >
+      <DraftStack.Screen name="DraftScreen" component={DraftScreen} />
+    </DraftStack.Navigator>
+  );
 }
