+++ src/screens/Authentication/Signout/SignoutScreen.tsx
@@ -3,7 +3,7 @@ import { View } from 'react-native';
 import { AuthContext } from '../../../context/AuthContext';
 import StyledButton from '../../../components/StyledButton/StyledButton';
 import styles from './styles';
-import { SignoutStackScreenProps } from '../../../navigation/types';
+import { SignoutStackScreenProps } from '../../../types/navigation';
 
 function SignoutScreen({
   navigation,
