+++ src/screens/Authentication/VerificationCode/VerificationCode.tsx
@@ -13,7 +13,7 @@ import { AuthContext } from '../../../context/AuthContext';
 import StyledButton from '../../../components/StyledButton/StyledButton';
 import logo from '../../../assets/cnsc-logo.png';
 import PhoneNumberScreen from '../PhoneNumber/PhoneNumber';
-import { AuthStackScreenProps } from '../../../navigation/types';
+import { AuthStackScreenProps } from '../../../types/navigation';
 
 function VerificationScreen({
   route,
