+++ src/screens/Feed/FeedScreen.tsx
@@ -7,7 +7,7 @@ import styles from './Styles';
 import { getAllJobs } from '../../firebase/firestore/job';
 import { Job } from '../../types/types';
 import Logo from '../../assets/cnsc-logo.png';
-import { FeedStackScreenProps } from '../../navigation/types';
+import { FeedStackScreenProps } from '../../types/navigation';
 
 function FeedScreen({ navigation }: FeedStackScreenProps<'FeedScreen'>) {
   const [open, setOpen] = useState(false);
